refactor: colocate deferred peel persistence helpers
