+++ crates/cgka-engine/src/message_processor/mod.rs
@@ -1181,97 +1181,6 @@ impl<S: StorageProvider> Engine<S> {
         Ok(normalization_pending.then_some(0).or(earliest))
     }
 
-    /// Normalize legacy/restarted lifecycle rows and persist at most one
-    /// sweep-sized slice. Returns whether additional normalized rows remain
-    /// to be persisted by a later scheduler tick.
-    fn normalize_deferred_peel_lifecycles(
-        &self,
-        records: &mut [cgka_traits::message::MessageRecord],
-        now: crate::convergence_clock::ConvergenceTime,
-    ) -> Result<bool, EngineError> {
-        let mut persisted = 0usize;
-        let mut normalization_pending = false;
-        for record in records {
-            let mut changed = false;
-            let lifecycle = record.deferred_peel.get_or_insert_with(|| {
-                changed = true;
-                cgka_traits::message::DeferredPeelLifecycle {
-                    first_observed_wall_ms: now.wall_ms,
-                    wall_high_water_ms: now.wall_ms,
-                    clock_instance_id: self.convergence_clock_instance_id,
-                    residence_deadline_monotonic_ms: now
-                        .monotonic_ms
-                        .saturating_add(self.deferred_peel_residence_ms),
-                    residence_deadline_wall_ms: now
-                        .wall_ms
-                        .saturating_add(self.deferred_peel_residence_ms),
-                    distinct_context_attempts: 0,
-                    last_context_fingerprint: None,
-                }
-            });
-            if lifecycle.clock_instance_id != self.convergence_clock_instance_id {
-                // Monotonic values do not survive restart. Preserve elapsed
-                // wall residence through a high-water mark. If wall time moved
-                // backwards, using the prior high water prevents premature
-                // expiry.
-                let observed_wall = lifecycle.wall_high_water_ms.max(now.wall_ms);
-                let remaining = lifecycle
-                    .residence_deadline_wall_ms
-                    .saturating_sub(observed_wall);
-                lifecycle.clock_instance_id = self.convergence_clock_instance_id;
-                lifecycle.residence_deadline_monotonic_ms =
-                    now.monotonic_ms.saturating_add(remaining);
-                lifecycle.wall_high_water_ms = observed_wall;
-                changed = true;
-            }
-            if changed {
-                if persisted < MAX_DEFERRED_ROWS_PER_SWEEP {
-                    self.storage.put_message(record)?;
-                    persisted += 1;
-                } else {
-                    normalization_pending = true;
-                }
-            }
-        }
-        Ok(normalization_pending)
-    }
-
-    fn release_deferred_peel_row(
-        &mut self,
-        record: &cgka_traits::message::MessageRecord,
-        resource: InboundResourceLimit,
-        disposition: crate::message_disposition::MessageDisposition,
-    ) -> Result<(), EngineError> {
-        let retry_count = record.deferred_peel.as_ref().map_or(0, |lifecycle| {
-            u64::from(lifecycle.distinct_context_attempts)
-        });
-        let residence_ms = record.deferred_peel.as_ref().map_or(0, |lifecycle| {
-            lifecycle
-                .wall_high_water_ms
-                .max(self.convergence_now().wall_ms)
-                .saturating_sub(lifecycle.first_observed_wall_ms)
-        });
-        self.storage.delete_message(&record.id)?;
-        self.audit_group(
-            &record.group_id,
-            crate::audit_helpers::deferred_peel_resource_refused_event(
-                hex::encode(record.id.as_slice()),
-                Some(record.epoch),
-                disposition.tag(),
-                retry_count,
-                residence_ms,
-            ),
-        );
-        self.events_buf
-            .push_back(GroupEvent::TransportObjectResourceRefused {
-                group_id: record.group_id.clone(),
-                message_id: record.id.clone(),
-                resource,
-            });
-        self.note_peel_deferred_row_retired(&record.group_id);
-        Ok(())
-    }
-
     /// Fingerprint of everything that can change a deferred peel's outcome:
     /// the group's live epoch and the retained peel-snapshot set. While this
     /// is unchanged, re-peeling a deferred row is guaranteed wasted work.
